% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spwss_stats.R
\name{get_spwss_stats}
\alias{get_spwss_stats}
\title{Calculate standardized pooled within-cluster sum of squares (SPWSS)}
\usage{
get_spwss_stats(dist_mat, labels)
}
\arguments{
\item{dist_mat}{a matrix of distance/dissimilarities on which to compute SPWSS}

\item{labels}{class labels}
}
\value{
spwss metric for specified class labels over distance matrix
}
\description{
Calculate standardized pooled within-cluster sum of squares (SPWSS)
}
\author{
Arshi Arora
}
