% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat-data.R, R/simsurvdat-data.R
\docType{data}
\name{simdat}
\alias{simdat}
\title{Simulated dataset with 3-class solution}
\format{
An object of class \code{"list"}

An object of class \code{"list"}
}
\usage{
data(simdat)

data(simdat)
}
\description{
A list of length 1 with a matrix simulated with 150 samples x 150 features with a 3-class structure
such that 15 features are distinct and associated with survival, other 15
features are just distinct and not associated with survival and remaining 120
are noise. See how this dataset was generated in the vignette

A matrix with simulate time-event data with 150 samples x 2 columns with a
3-class structure with median survival of 4.5, 3.25 and 2 yrs respectively.
such that 15 features are distinct and associated with survival, other 15
features are just distinct and not associated with survival and remaining 120
are noise. See how this dataset was generated in the vignette
}
\examples{
data(simdat)
class(simdat)
dim(simdat[[1]])
simdat[[1]][1:5,1:5]
data(simdat)
class(simdat)
dim(simdat[[1]])
simdat[[1]][1:5,1:5]
}
\keyword{datasets}
