% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvm_dat-data.R
\docType{data}
\name{uvm_dat}
\alias{uvm_dat}
\title{TCGA UVM Mutation and Copy Number datasets}
\format{
An object of class \code{"list"}
}
\usage{
data(uvm_dat)
}
\description{
A list of length 2 with
TCGA UVM Mutation data with 80 samples and 87 genes
TCGA UVM Copy Number data with 80 samples and 749 segments. See Appendix
in vignette for more details. Teh data is downloaded from here
https://gdc.cancer.gov/about-data/publications/pancanatlas
}
\examples{
data(uvm_dat)
uvm_dat[[1]][1:5,1:5]
uvm_dat[[2]][1:5,1:5]
}
\keyword{datasets}
