% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvm_survClust_cv.fit-data.R
\docType{data}
\name{uvm_survClust_cv.fit}
\alias{uvm_survClust_cv.fit}
\title{survClust cv.survclust output of integrated TCGA UVM Mutation and Copy Number datasets.}
\format{
An object of class \code{"list"}
}
\usage{
data(uvm_survClust_cv.fit)
}
\description{
The output is a list object consisting of 6 sub-lists for \code{k = 2:7}, with 10 \code{cv.survclust} outputs (for each round of cross-validation), each consisting of \code{cv.labels, cv.logrank, cv.spwss} for 3 folds.
}
\examples{
data(uvm_survClust_cv.fit)
names(uvm_survClust_cv.fit[[1]][[1]])
}
\keyword{datasets}
