% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvm_survdat-data.R
\docType{data}
\name{uvm_survdat}
\alias{uvm_survdat}
\title{TCGA UVM Clinical file}
\format{
An object of class \code{"matrix"}
}
\usage{
data(uvm_survdat)
}
\description{
A matrix with 2 columns, with first column as OS.time and second
column as OS events. The data is downloaded from - https://gdc.cancer.gov/about-data/publications/pancanatlas
}
\examples{
data(uvm_survdat)
head(uvm_survdat)
}
\keyword{datasets}
