% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survClust.R
\name{survClust}
\alias{survClust}
\title{perform supervised clustering for a particular \code{k}}
\usage{
survClust(combine.dist, survdat, k, cmd.k = NULL)
}
\arguments{
\item{combine.dist}{integrated weighted distance matrix from \code{combineDist}}

\item{survdat}{A nx2 matrix consisting of survival data with \code{n} samples and first column as time and second column as events, with samples as rownames}

\item{k}{choice of \code{k} to perform clustering on samples}

\item{cmd.k}{number of dimensions used by \code{cmdscale} to perform clustering on samples. Defaults is \code{n-1}}
}
\value{
\itemize{
 \item{fit} { returns a list , \code{fit} consisting of all clustering samples as in \code{kmeans} 
 \code{fit.lr}, computed logrank statistic between \code{k} clusters}
}
}
\description{
\code{survClust} function performs supervised clustering on a \code{combineDist} output for a particular \code{k}. 
It uses all \code{n-1} dimensions for clustering.
}
\author{
Arshi Arora
}
